+++ lib/custom_chewie/src/video_view.dart
@@ -1,5 +1,5 @@
 import 'package:chewie_with_danmaku/application_controller.dart';
-import 'package:chewie_with_danmaku/src/flutter_danmaku_controller.dart';
+import 'package:chewie_with_danmaku/danmaku_bullet/bullet_setting.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:get/get.dart';
@@ -9,29 +9,70 @@ import 'package:video_player/video_player.dart';
 import '../../danmaku_bullet/reusable_component.dart';
 import '../chewie.dart';
 
-///播放器组件
 class VideoView extends StatefulWidget {
-  final String? progressBarIndicatorImagePath;
+  /// The URL that contains the video to be played
   final String url;
+
+  /// If video will start playing upon opening the page/using the widget
   final bool autoPlay;
+
+  /// If the video will auto loop itself
   final bool looping;
+
+  /// Specify the aspect ratio of the given video
   final double aspectRatio;
-  final Function(FlutterDanmakuController) onControllerInitialized;
+
+  /// A list of button that can be added to the right of the screen
+  /// when the video is in full screen mode
   final List<Widget>? rightButtonList;
+
+  /// The color used on progress bar
   final ChewieProgressColors? progressColors;
+
+  /// The title of the video that will display on top of video
   final String? videoTitle;
+
+  /// The image path to change the video progress bar icon
+  final String? progressBarIndicatorImagePath;
+
   const VideoView(
     this.url, {
     Key? key,
     this.autoPlay = false,
     this.looping = false,
     this.aspectRatio = 16 / 9,
-    required this.onControllerInitialized,
     this.rightButtonList,
     this.progressColors,
     this.videoTitle,
     this.progressBarIndicatorImagePath,
   }) : super(key: key);
+
+  static Widget createWithDependencies(
+    String url, {
+    bool autoPlay = false,
+    bool looping = false,
+    double aspectRatio = 16 / 9,
+    List<Widget>? rightButtonList,
+    ChewieProgressColors? progressColors,
+    String? videoTitle,
+    String? progressBarIndicatorImagePath,
+  }) {
+    // Initialize dependencies here
+    Get.put(ApplicationController());
+    Get.put(BulletController());
+
+    // Return an instance of VideoView
+    return VideoView(
+      url,
+      autoPlay: autoPlay,
+      looping: looping,
+      aspectRatio: aspectRatio,
+      rightButtonList: rightButtonList,
+      progressColors: progressColors,
+      videoTitle: videoTitle,
+      progressBarIndicatorImagePath: progressBarIndicatorImagePath,
+    );
+  }
 
   @override
   _VideoViewState createState() => _VideoViewState();
@@ -40,7 +81,7 @@ class VideoView extends StatefulWidget {
 class _VideoViewState extends State<VideoView> {
   ApplicationController controller = Get.find<ApplicationController>();
 
-  /// 进度条颜色配置
+  /// Default video progress bar color
   get _progressColors => ChewieProgressColors(
       playedColor: Colors.pinkAccent,
       handleColor: Colors.pinkAccent,
@@ -50,7 +91,6 @@ class _VideoViewState extends State<VideoView> {
   @override
   void initState() {
     super.initState();
-    //初始化播放器设置
     controller.videoPlayerController =
         VideoPlayerController.networkUrl(Uri.parse(widget.url));
     controller.chewieController = ChewieController(
@@ -66,7 +106,8 @@ class _VideoViewState extends State<VideoView> {
         progressBarIndicatorImagePath: widget.progressBarIndicatorImagePath,
         videoTitle: widget.videoTitle,
         rightButtonList: widget.rightButtonList,
-        onControllerInitialized: widget.onControllerInitialized,
+        onControllerInitialized:
+            Get.find<ApplicationController>().setController,
         bottomGradient: blackLinearGradient(),
         aspectRatio: widget.aspectRatio,
       ),