+++ lib/main.dart
@@ -1,12 +1,9 @@
 import 'package:chewie_with_danmaku/custom_chewie/src/video_view.dart';
-import 'package:chewie_with_danmaku/danmaku_bullet/bullet_setting.dart';
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-import 'application_controller.dart';
 
 void main() {
-  Get.put(ApplicationController());
-  Get.put(BulletController());
+  // Get.put(ApplicationController());
+  // Get.put(BulletController());
   runApp(const MyApp());
 }
 
@@ -36,12 +33,12 @@ class MyHomePage extends StatefulWidget {
 class _MyHomePageState extends State<MyHomePage> {
   @override
   Widget build(BuildContext context) {
+    // Get.put(ApplicationController());
+    // Get.put(BulletController());
     return Scaffold(
       body: Center(
-          child: VideoView(
+          child: VideoView.createWithDependencies(
         "https://flutter.github.io/assets-for-api-docs/assets/videos/butterfly.mp4",
-        onControllerInitialized:
-            Get.find<ApplicationController>().setController,
         autoPlay: true,
       )),
     );