+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 1.0.0
+
 Initial version of chewie_with_danmaku
+
 -use a chewie controller to play video
+
 -have a initial customized UI
+
 -have danmaku function included
+
+1.0.1
+
+Changed some text and added comments to some function
+
+1.0.2
+
+Updated the call of getx controller to make the widget ready to use
+
+1.0.3
+
+Added assets file to be displayed as default